refactor: CoupleRepository를 couple 도메인 패키지로 이동
- goal.repository → couple.repository로 이동
- 도메인 레이어 분리 개선
- CoupleService에서 정규화된 import 사용
